Refactor in progress.
Pull out Kinesis stuff into helper file.
