+++ example/spec/kinesisTests/KinesisTestTriggerSpec.js
@@ -1,24 +1,21 @@
 // npm packages
-const _ = require('lodash');
 const fs = require('fs');
 const Handlebars = require('handlebars');
-const { Kinesis, StepFunctions, S3 } = require('aws-sdk');
+const { StepFunctions, S3 } = require('aws-sdk');
 
 jasmine.DEFAULT_TIMEOUT_INTERVAL = 550000;
 
 const {
-  getWorkflowArn,
   LambdaStep,
-  timeout,
   waitForCompletedExecution
 } = require('@cumulus/integration-tests');
 const { randomString } = require('@cumulus/common/test-utils');
 
 const { loadConfig } = require('../helpers/testUtils');
+const { createNewTestStream, putRecordOnStream, waitForTestSfStarted } = require('../helpers/kinesisHelpers');
 const testConfig = loadConfig();
 const lambdaStep = new LambdaStep();
-const kinesis = new Kinesis({ region: testConfig.awsRegion });
-const sfn = new StepFunctions({ region: testConfig.awsRegion });
+
 const s3 = new S3({ region: testConfig.awsRegion });
 const streamName = testConfig.streamName;
 const granuleId = 'L2_HR_PIXC_product_0001-of-4154';
@@ -66,26 +63,10 @@ const expectedSyncGranulesPayload = {
 };
 
 
-async function getLastExecution() {
-  const kinesisTriggerTestStpFnArn = await getWorkflowArn(testConfig.stackName, testConfig.bucketName, 'KinesisTriggerTest');
-  return new Promise((resolve, reject) => {
-    sfn.listExecutions({ stateMachineArn: kinesisTriggerTestStpFnArn }, (err, data) => {
-      if (err) {
-        reject(err);
-      }
-      else {
-        resolve(_.orderBy(data.executions, 'startDate', 'desc')[0]);
-      }
-    });
-  });
-}
+// When Cumulus is configured to trigger a CNM workflow from a Kinesis stream and a message appears on the stream, Cumulus triggers the workflow
 
-describe('The Ingest Kinesis workflow', () => {
-  const waitTimeInterval = 1000;
+describe('The Cloud Notification Mechanism Kinesis workflow', () => {
   const maxWaitTime = 240000;
-  let timeWaited = 0;
-  let putRecord;
-  let lastExecution;
   let workflowExecution;
   let executionStatus;
   let s3FileHead;
@@ -93,69 +74,19 @@ describe('The Ingest Kinesis workflow', () => {
   beforeAll(async () => {
     // Create stream. REVIEW: Do we really need this? Developers may likely
     // setting up the stream themselves, once.
-    await new Promise((resolve, reject) => {
-      kinesis.describeStream({ StreamName: streamName }, (err, data) => {
-        if (err && err.code === 'ResourceNotFoundException') {
-          kinesis.createStream({ StreamName: streamName, ShardCount: 1 }, (err, data) => {
-            if (err) reject(err);
-            resolve(data);
-          });
-        }
-        else if (err) {
-          reject(err);
-        }
-        else {
-          resolve(data);
-        }
-      });
-    });
 
-
-    putRecord = await new Promise((resolve, reject) => {
-      kinesis.putRecord({
-        Data: JSON.stringify(record),
-        PartitionKey: '1',
-        StreamName: streamName
-      }, (err, data) => {
-        if (err) reject(err);
-        resolve(data);
-      });
-    });
-
-
-    // Wait until a we discover an execution has started which matches our record identifier.
-    // That will identify the execution we want to test.
-    while (timeWaited < maxWaitTime && workflowExecution === undefined) {
-      try {
-        await timeout(waitTimeInterval);
-        timeWaited += waitTimeInterval;
-        try {
-          lastExecution = await getLastExecution();
-        }
-        catch (error) {
-          console.log(error);
-          throw new Error(error);
-        }
-        // getLastExecution returns undefined if no previous execution exists
-        if (lastExecution && lastExecution.executionArn) {
-          try {
-            const taskOutput = await lambdaStep.getStepOutput(lastExecution.executionArn, 'sf2snsStart');
-            if (taskOutput.payload.identifier === recordIdentifier) {
-              workflowExecution = lastExecution;
-            }
-          }
-          catch (error) {
-            console.log(error);
-            throw new Error(error);
-          }
-        }
-      }
-      catch (error) {
-        console.log(error);
-        throw new Error(error);
-      }
+    try {
+      await createNewTestStream(streamName);
+      console.log('createNewTestStream');
+      await putRecordOnStream(streamName, record);
+      console.log('putRecordOnStream');
+      workflowExecution = await waitForTestSfStarted(recordIdentifier, maxWaitTime);
+      console.log('workflow running', workflowExecution);
     }
-
+    catch (e) {
+      console.log(e);
+      throw new Error(e);
+    }
     // Wait for our execution to complete so we can test the outputs.
     if (workflowExecution === undefined) {
       throw new Error('Timeout waiting for new execution to start');
